+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Add `:article` selector and `have_article` matcher
 - Add `:grid`, `:gridcell`, `:columnheader`, and `:row`
 - Add `:menu` and `:menuitem` selector
 - Add `current:` filter for `:link` and `:link_or_button` selectors